+++ packages/assets/src/duotones/Gift.tsx
@@ -1,10 +1,9 @@
 import React from 'react';
 import styled from 'styled-components';
-import { useTheme } from '@heathmont/moon-themes';
+import { ColorProps, useTheme } from '@heathmont/moon-themes';
 import { themed } from '@heathmont/moon-utils';
-import SvgProps from './private/SvgProps';
 
-const Svg = ({ secondaryColor, ...rest }: SvgProps) => (
+const Svg = ({ secondaryColor, ...rest }: any) => (
   <svg
     width="1em"
     height="1em"
@@ -122,6 +121,14 @@ const Svg = ({ secondaryColor, ...rest }: SvgProps) => (
   </svg>
 );
 
+type SvgProps = {
+  color?: ColorProps;
+  height?: string | number;
+  width?: string | number;
+  fontSize?: string | number;
+  verticalAlign?: string;
+};
+
 const Component = styled(Svg)<SvgProps>(
   ({ color, height, width, fontSize, verticalAlign, theme }) => ({
     ...(color && {
@@ -134,11 +141,9 @@ const Component = styled(Svg)<SvgProps>(
   })
 );
 
-const Gift: React.FC<SvgProps> = (props) => {
+const Gift: React.FC<any> = (props) => {
   const theme = useTheme();
-  const secondaryColor = theme.color.bulma[100];
-  const attributes = { ...props, secondaryColor } as any;
-  return <Component {...attributes} />;
+  return <Component secondaryColor={theme.color.bulma[100]} {...props} />;
 };
 
 Gift.defaultProps = {