+++ packages/themes/src/sharedTokens/sharedTokens.ts
@@ -1,4 +1,3 @@
-import { CSSObject } from 'styled-components';
 import type { ColorShared, ColorValue } from '../supportColors/supportColors';
 
 /**
@@ -331,7 +330,7 @@ export interface SharedTheme {
   borderStyle: BorderStyle;
   boxShadow: BoxShadow; // legacy
   breakpoint: Breakpoint;
-  fontFace?: CSSObject[];
+  fontFace?: any;
   fontFamily: FontFamily;
   fontSize: FontSize;
   fontWeight: FontWeight;