+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 7.15.1
+
+### Patch Changes
+
+- Revert assets ts fixes & Popover min & max width
+- Updated dependencies [undefined]
+  - @heathmont/moon-themes@7.15.1
+
 ## 7.15.0
 
 ### Minor Changes